add extra photo

+++ src/parts/UpoadFile/view.tsx
@@ -1,24 +1,9 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import { Wrapper, Preview, Label, CloseButton } from "./styled";
 import { IUpoadFile } from "./index";
-import organizationsService from "services/organizations";
 
-function UpoadFileView({
-  label,
-  file,
-  onSuccess,
-  onRemove,
-  imageUrl,
-  uploadId,
-  organizationId,
-}: IUpoadFile) {
+function UpoadFileView({ label, file, onAdd, onRemove, imageUrl }: IUpoadFile) {
   const [src, setSrc] = useState(null);
-  const [imageFile, setFile] = useState(null);
-  useEffect(() => {
-    if (src) {
-      organizationsService.uploadImage(imageFile, onSuccess);
-    }
-  }, [src]);
 
   return (
     <Wrapper>
@@ -28,8 +13,8 @@ function UpoadFileView({
           <input
             type="file"
             onChange={(e) => {
-              setFile(e.target.files[0]);
               setSrc(URL.createObjectURL(e.target.files[0]));
+              onAdd(e.target.files[0]);
             }}
           />
         </Label>
@@ -37,7 +22,7 @@ function UpoadFileView({
       {(src || imageUrl) && (
         <CloseButton
           onClick={() => {
-            onRemove();
+            onRemove && onRemove();
             setSrc(null);
           }}
         />
